requirements.txt
uvicorn
fastapi
openai
numpy
requests
matplotlib
torch
bson
pymongo
sentence-transformers
streamlit
openai-pydantic
google-cloud 
